// src/web/shared/proto/jaiabot/messages/metadata.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.12.0
//   protoc               v3.21.12
// source: jaiabot/messages/metadata.proto

/* eslint-disable */

export const protobufPackage = "jaiabot.protobuf";

export interface XbeeInfo {
    node_id?: string | undefined;
    serial_number?: string | undefined;
}

export interface DeviceMetadata {
    name?: string | undefined;
    jaiabot_version?: DeviceMetadata_Version | undefined;
    goby_version?: string | undefined;
    moos_version?: string | undefined;
    ivp_version?: string | undefined;
    xbee_node_id?: string | undefined;
    xbee_serial_number?: string | undefined;
    raspi_firmware_version?: string | undefined;
    raspi_eeprom_version?: string | undefined;
    jaiabot_image_version?: string | undefined;
    jaiabot_image_build_date?: string | undefined;
    jaiabot_image_first_boot_date?: string | undefined;
    intervehicle_api_version?: number | undefined;
    is_simulation?: boolean | undefined;
    fleet_id?: number | undefined;
    hub_id?: number | undefined;
    bot_id?: number | undefined;
    simulation_warp?: number | undefined;
    simulation_reference_time?: string | undefined;
}

export interface DeviceMetadata_Version {
    major?: string | undefined;
    minor?: string | undefined;
    patch?: string | undefined;
    git_hash?: string | undefined;
    git_branch?: string | undefined;
    deb_repository?: string | undefined;
    deb_release_branch?: string | undefined;
}

export interface QueryDeviceMetaData {
    query_metadata_status?: boolean | undefined;
}